Refactoring annotation processor: extract pattern mapper
